# backend/swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  description: API documentation for the E-commerce backend.
  version: 3.0.0
servers:
  - url: http://localhost:3000
    description: Local development server

#  user api documentation
paths:
...
  /category/update/{id}:
    put:
      summary: Admin Update category
      description: Update the details of an existing category by ID.
      tags:
        - Category
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          description: category Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the category.
                status:
                  type: string
                  description: The status of the category (e.g., active, inactive).
              required:
                - name
                - status
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  name: string
        "500":
          description: Internal server error

  # category list get api documentation
  /category/list:
    get:
      summary: Get all categories
      description: Retrieve a list of all category.
      tags:
        - Category
      operationId: categoryList
      responses:
        "200":
          description: Successfully retrieved category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  categoryList:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: The unique identifier of the category.
                        name:
                          type: string
                          description: The name of the category.
                        status:
                          type: string
                          description: The status of the category .
                        isDeleted:
                          type: boolean
                          description: Whether the category has been deleted .
        "500":
          description: Internal server error

  # category delete api documentation
  /category/delete/{id}:
    delete:
      summary: Admin Delete category
      description: category as deleted using its Id.
      tags:
        - Category
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          description: category Id
          schema:
            type: string
      responses:
        "200":
          description: Category deleted Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Category deleted successfully"
        "400":
          description: Bad request
        "404":
          description: Category not found
        "500":
          description: Internal server error

  # subCategory  post api documentation
...
  /product/update/{id}:
    put:
      summary: Admin Update product
      description: update product details by existing product Id.
      tags:
        - Product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          description: product Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                  description: category Id
                  example: "60d21b4667d0d8992e610c85"
                subCategoryId:
                  type: string
                  description: subCategory Id
                  example: "60d21b4667d0d8992e610c85"
                name:
                  type: string
                  example: "sony TV"
                image:
                  type: string
                  example: "https://example.com/sonyIV.jpg"
                description:
                  type: string
                  example: "OlED SMART TV"
                quantity:
                  type: Number
                  example: 1
                price:
                  type: Number
                  example: 1000
                status:
                  type: string
                  enum: [active, inactive]
                  default: "active"
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updateProduct:
                    type: object
        "500":
          description: Internal server error

  # product delete api documentation
  /product/delete/{id}:
    delete:
      summary: Admin Deleted product
      description: Delete product by product Id.
      tags:
        - Product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          description: product Id
          schema:
            type: string
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product deleted successfully
        "500":
          description: Internal server error

  #product list get api documentation
...
  /cart/delete/{id}:
    delete:
      summary: Cart delete by Id
      tags:
        - Cart
      operationId: deleteCart
      parameters:
        - name: id
          in: path
          required: true
          description: cart Id
          schema:
            type: string
      responses:
        "200":
          description: Product remove form cart
        "500":
          description: Internal server error

  # wishlist  post api documentation
...
  /discount/delete/{id}:
    delete:
      summary: delete discount
      tags:
        - Discount
      operationId: deleteDiscount
      parameters:
        - name: id
          in: path
          required: true
          description: discount Id
          schema:
            type: string
      responses:
        "200":
          description: discount deleted successfully
        "500":
          description: Internal server error

  # discount get api documentation
...